Check import rules in tsx files
- Enable "import/extensions"
- Break import cycle in Components
- Fix App import name
