+++ frontend/src/components/widgets/CustomComponent/ComponentInstance.tsx
@@ -26,6 +26,7 @@ import { Source, WidgetStateManager } from "lib/WidgetStateManager"
 import React, { createRef, ReactNode } from "react"
 import queryString from "query-string"
 import { ComponentRegistry } from "./ComponentRegistry"
+import { ComponentMessageType, StreamlitMessageType } from "./enums"
 
 /**
  * The current custom component API version. If our API changes,
@@ -33,30 +34,6 @@ import { ComponentRegistry } from "./ComponentRegistry"
  * version in the COMPONENT_READY call.
  */
 export const CUSTOM_COMPONENT_API_VERSION = 1
-
-/** Messages from Component -> Streamlit */
-export enum ComponentMessageType {
-  // A component sends this message when it's ready to receive messages
-  // from Streamlit. Streamlit won't send any messages until it gets this.
-  // Data: { apiVersion: number }
-  COMPONENT_READY = "streamlit:componentReady",
-
-  // The component has a new value. Send it back to Streamlit, which
-  // will then re-run the app.
-  // Data: { value: any }
-  SET_COMPONENT_VALUE = "streamlit:setComponentValue",
-
-  // The component has a new height for its iframe.
-  // Data: { height: number }
-  SET_FRAME_HEIGHT = "streamlit:setFrameHeight",
-}
-
-/** Messages from Streamlit -> Component */
-export enum StreamlitMessageType {
-  // Sent by Streamlit when the component should re-render.
-  // Data: { args: any, disabled: boolean }
-  RENDER = "streamlit:render",
-}
 
 interface Props {
   registry: ComponentRegistry
